+++ sec_gov_edgar_client/__init__.py
@@ -1,10 +1,12 @@
 from .central_index_key import CIKRepositoryInterface, LocalFileCIKRepository
-from .client import BalanceSnapshot, Report, Reports, SECGovEDGARClient
+from .client import SECGovEDGARClient, UserAgent
+from .parser import BalanceSnapshot, Reports
 
 __all__ = (
     "SECGovEDGARClient",
+    "UserAgent",
+
     "Reports",
-    "Report",
     "BalanceSnapshot",
 
     "CIKRepositoryInterface",