+++ sec_gov_edgar_client/client.py
@@ -1,5 +1,5 @@
-from datetime import date
-from typing import Iterable, Mapping, Optional
+from functools import cached_property
+from typing import Optional
 
 import attr
 import ujson
@@ -7,29 +7,23 @@ from aiohttp import ClientSession
 from yarl import URL
 
 from .central_index_key import CIKRepositoryInterface
+from .parser import Reports, SECResponseParser
 
 __all__ = (
-    "Reports",
-    "Report",
-    "BalanceSnapshot",
     "SECGovEDGARClient",
+    "UserAgent",
 )
 
 
-@attr.s(auto_attribs=True, slots=True, frozen=True)
-class BalanceSnapshot:
-    assets: int
-    stockholders_equity: int
+@attr.s(auto_attribs=True, frozen=True)
+class UserAgent:
+    company: str
+    user: str
+    email: str
 
-
-@attr.s(auto_attribs=True, slots=True, frozen=True)
-class Report:
-    balance: Mapping[date, BalanceSnapshot]
-
-
-@attr.s(auto_attribs=True, slots=True, frozen=True)
-class Reports:
-    annual: Report
+    @cached_property
+    def view(self) -> str:
+        return f"{self.company} {self.user} <{self.email}>"
 
 
 @attr.s(auto_attribs=True, slots=True, frozen=True)
@@ -37,6 +31,11 @@ class SECGovEDGARClient:
     _session: ClientSession
     _url: URL
     _CIKs: CIKRepositoryInterface
+    _user_agent: UserAgent = UserAgent(
+        "Tribe.invest",
+        "Danila Korobkov",
+        "korobkov.danila.yurevich@gmail.com",
+    )
 
     async def get_reports(
         self,
@@ -51,61 +50,15 @@ class SECGovEDGARClient:
         self,
         cik: str,
     ) -> Reports:
-        url = self._url.with_path(f"xbrl/companyfacts/CIK{cik}.json")
-        async with self._session.get(url) as response:
+        url = self._url / f"xbrl/companyfacts/CIK{cik}.json"
+        async with self._session.get(
+            url,
+            headers=self._get_headers(),
+        ) as response:
             data = await response.json(loads=ujson.loads)
+            return SECResponseParser(data).parse_reports()
 
-            assets = _get_statements(data, key="Assets")
-            stockholders_equity = _get_statements(
-                data,
-                key="StockholdersEquity",
-            )
-            stockholders_equity = {
-                key: value
-                for key, value in stockholders_equity.items()
-                if key in assets
-            }
-
-            balance = {
-                reported_at: BalanceSnapshot(
-                    asset,
-                    stockholders_equity[reported_at],
-                )
-                for reported_at, asset in assets.items()
-            }
-            return Reports(
-                annual=Report(balance),
-            )
-
-
-def _get_statements(data: dict, key: str) -> dict[date, int]:
-    statements = data["facts"]["us-gaap"][key]["units"]["USD"]
-    sorted_statements = _sort_by_period_end(statements)
-    annual_statements = list(_filter_annual_statements(sorted_statements))
-    return _get_date_value(annual_statements)
-
-
-def _sort_by_period_end(statements: Iterable[dict]) -> Iterable[dict]:
-    return sorted(
-        statements,
-        key=lambda statement: (
-            date.fromisoformat(statement["end"]), statement["fy"],
-        ),
-    )
-
-
-def _filter_annual_statements(statements: Iterable[dict]) -> Iterable[dict]:
-    return (
-        statement
-        for statement in statements
-        if statement["form"] == "10-K" and "frame" not in statement
-    )
-
-
-def _get_date_value(statements: Iterable[dict]) -> dict[date, int]:
-    res = {}
-    for statement in statements:
-        end = date.fromisoformat(statement["end"])
-        if end not in res:
-            res[end] = statement["val"]
-    return res
+    def _get_headers(self) -> dict[str, str]:
+        return {
+            "User-Agent": self._user_agent.view,
+        }